pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "brentford-onsale-calendar"
version = "0.1.0"
description = "Sync Brentford FC ticket on-sale dates to Google Calendar"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "click>=8.3.0",
    "google-auth>=2.34.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.150.0",
    "html5lib>=1.1",
    "pydantic>=2.0.0",
    "python-dateutil>=2.9.0",
    "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "pre-commit>=4.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-beautifulsoup4>=4.12.0",
    "types-html5lib>=1.1.0",
    "types-requests>=2.32.0",
    "google-api-python-client-stubs>=1.30.0",
]

[project.scripts]
brentford-calendar = "brentford_calendar.cli:main"

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
]

[tool.ruff.lint.isort]
known-first-party = ["brentford_calendar"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"

[tool.hatch.build.targets.wheel]
packages = ["src/brentford_calendar"]
